feat: Add type definitions and document processing utility
- Added type definitions for  in .
- Added type definitions for  in .
- Implemented  utility in  to handle processing of , , , and  files.
- Configured PDF.js worker to be served from  for Vite

Modified textareas of dashboard and notepreview to both look and work thesame, fixed some buttons and colouring inconsistencies, removed the start test button from the lbrary page into the textarea for a more logical flow, added a logic to regenerate questions on source modifcation.

Tried implementing a pdf analyzing library so it can be parsed into text, but it did not go well, i'll have to revisit that later.